rename toast
